requirements.txt
altgraph==0.17
et-xmlfile==1.0.1
future==0.18.2
labjack-ljm==1.21.0
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
pefile==2019.4.18
Pillow==8.2.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
six==1.15.0
